Tests split for performance
