+++ src/main/scala/math/cat/topos/ToposLogic.scala
@@ -1,12 +1,17 @@
 package math.cat.topos
 
-import scala.language.postfixOps
 import math.cat.SetFunction
 import math.cat.topos.CategoryOfDiagrams.DiagramArrow
 import math.sets.Sets.{set, setOf}
 import scalakittens.Result
 
+import scala.collection.mutable
+import scala.language.postfixOps
+
 trait ToposLogic { topos: CategoryOfDiagrams =>
+
+// TODO: use the right tagging, find the right predicate
+//  private[this] val cache = mutable.Map[(String, Predicate, Predicate), Predicate]()
 
   trait Predicate extends DiagramArrow { p: DiagramArrow =>
     val d1: Diagram = Ω
@@ -16,7 +21,7 @@ trait ToposLogic { topos: CategoryOfDiagrams =>
       if (ts.contains("∧") || ts.contains("∨") || ts.contains("=>")) s"($ts)" else ts
     }
 
-    private def tag2(tag1: Any, op: String, tag2: Any): String = s"${wrapTag(tag1)} $op ${wrapTag(tag1)}"
+    private def tag2(tag1: Any, op: String, tag2: Any): String = s"${wrapTag(tag1)} $op ${wrapTag(tag2)}"
 
     private def setAt(o: Any): set = {
       val o1 = domainCategory.obj(o)
@@ -27,19 +32,21 @@ trait ToposLogic { topos: CategoryOfDiagrams =>
     private def transformAt(o: Any): SetFunction =
       transformPerObject(domainCategory.obj(o)).asInstanceOf[SetFunction]
 
-    private def binaryOp(q: Predicate, opTag: String, ΩxΩ_to_Ω: DiagramArrow): Predicate = {
-      binaryOpNamed(q, ΩxΩ_to_Ω, tag2(p.tag, opTag, q.tag))
+    private def binaryOp(q: Predicate, opName: String, ΩxΩ_to_Ω: DiagramArrow): Predicate = {
+      binaryOpNamed(q, ΩxΩ_to_Ω, tag2(p.tag, opName, q.tag))
     }
 
-    private[ToposLogic] def binaryOpNamed(q: Predicate, ΩxΩ_to_Ω: DiagramArrow, newTag: String): Predicate = {
-      require(q.d0 == p.d0)
-      
-      for (o <- domainCategory.objects) {
-        require(p.setAt(o) == q.setAt(o), s"Different domains at $o for ${p.tag} and ${q.tag}")
-      }
+    private[ToposLogic] def binaryOpNamed(q: Predicate, ΩxΩ_to_Ω: DiagramArrow, name: String): Predicate =
+    // TODO: when identification is fixed (finding the right point), uncomment
+    // cache.getOrElseUpdate((name, p, q), 
+        evalBinaryOp(q, ΩxΩ_to_Ω, name)
+    //)
+    
+    private[ToposLogic] def evalBinaryOp(q: Predicate, ΩxΩ_to_Ω: DiagramArrow, newTag: String): Predicate = {
+      requireCompatibility(q)
 
       new Predicate {
-        val d0: Diagram = p.d0
+        val d0: Obj = p.d0
         val tag: Any = newTag
 
         override def transformPerObject(o: domainCategory.Obj): codomainCategory.Arrow = {
@@ -48,7 +55,6 @@ trait ToposLogic { topos: CategoryOfDiagrams =>
           val po = p.transformAt(o)
           val qo = q.transformAt(o)
 
-          def trans(v: Any): Any = (po(v), qo(v))
           val PQtoΩxΩ: SetFunction =
             new SetFunction(
               s"PQ->ΩxΩ($o)",
@@ -56,7 +62,6 @@ trait ToposLogic { topos: CategoryOfDiagrams =>
               v => (po(v), qo(v))
             )
 
-          val conj: SetFunction = Ω.conjunction(o).asInstanceOf[SetFunction]
           val op: SetFunction = ΩxΩ_to_Ω(o).asInstanceOf[SetFunction]
           val maybeFunction = PQtoΩxΩ andThen op
           codomainCategory.arrow(Result(maybeFunction).iHope)
@@ -64,8 +69,17 @@ trait ToposLogic { topos: CategoryOfDiagrams =>
       }
     }
 
+    private def requireCompatibility(q: Predicate): Unit = {
+      require(q.d0 == p.d0)
+
+      for (o <- domainCategory.objects) {
+        require(p.setAt(o) == q.setAt(o), s"Different domains at $o for ${p.tag} and ${q.tag}")
+      }
+    }
+
     /**
       * Conjunction with another predicate
+      *
       * @param q another predicate
       * @return their conjunction
       */
