+++ src/test/scala/math/cat/topos/Fixtures.scala
@@ -1,6 +1,8 @@
 package math.cat.topos
 
 import math.Test
+import math.cat.Categories.{Cat, SomeKnownCategories}
+import math.cat.Category
 import org.specs2.matcher.MatchResult
 import org.specs2.execute.{Result => TestResult}
 import scalakittens.Result
@@ -8,6 +10,9 @@ import scalakittens.Result._
 
 class Fixtures extends Test with TestDiagrams {
   type SUT = Diagram
+
+  def report(cat: Category)(what: String): Unit =
+    println(s"  checking $what over ${cat.name}")
   
   def expectOk(r: Result[_]): TestResult = {
     r.isGood aka r.toString must beTrue
@@ -29,12 +34,13 @@ class Fixtures extends Test with TestDiagrams {
       unit: P,
       binop: (P, P) => P): MatchResult[Any] = {
       import topos._
+      val rep = report(topos.domain) _
       val points = Ω.points
       println(s"Testing <<${domain.name}>> $what monoidal properties (${points.size} points in Ω)")
       def predicate(p: Point): P = p.asPredicate.asInstanceOf[P]
 
       for {pt1 <- points } {
-        println(s"  monoidal at ${pt1.tag}")
+        rep(s"monoidal at ${pt1.tag}")
         val p = predicate(pt1)
         binop(unit, p) === p
         // idempotence
@@ -56,6 +62,17 @@ class Fixtures extends Test with TestDiagrams {
       }
       ok
     }
+  }
+
+  val categoriesToTest: List[Cat] = SomeKnownCategories
+
+  val batchSize = 8
+
+  val groupedCategoriesToTest: List[List[Cat]] = {
+    (for {
+      i <- 0 until batchSize
+      indices = i until SomeKnownCategories.length by batchSize
+    } yield indices map (SomeKnownCategories(_)) toList) toList
   }
 
 }
